project1/cloud-functions/requirements.txt
google.cloud
google-api-core
google-cloud-storage
google-cloud-datastore
google-api-python-client
google-cloud-pubsub
oauth2client